Fix build issue caused by misnamed test file
Presently, builds are broken because the utils_test.ts file in
api/src/folder/controller imports a function from the
utilities in the /api/test directory. As a result tsc gets confused
regarding what the root folder of the API package is, causing it to
output in an unexpected place. This causes the Dockerfile to be unable
to find the index.js needed to start the API server. This commit fixes
this problem by moving the utilities from utils_test.ts to the test
files that rely on them. This causes some duplication, but
1. I don't see an elegant way to get tsc to ignore a controller-specific
   test utils file without vitest becoming convinced that there's a test
   file with no tests in it (which it sees as an error)
2. The utils in utils_test.ts were for loading an clearing fixture data.
   It is probably best not to reuse functions for this across test
   modules, because these are expensive operations in terms of time and
   reusing them encourages us to load more data than a particular module
   might need.
